app/game-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
typing-extensions==4.8.0